Console: check printf formats, and fix the 23 bugs that finds
IConsoleWriter::WriteLn/Error/Warning/Debug are varargs functions
taking a format string, and nothing verified the two against each
other.  That is how a std::string_view reached a %s in IsoFile::open()
and segfaulted on every disc the reader could not get through.  Adding
__attribute__((format(printf, 2, 3))) - member functions count `this'
as argument one - turns that whole class into a compile error, and the
compiler immediately found 23 more instances across 21 sites.

The one that matters most is not a crash but a format string an
attacker controls:

    const std::string line(file.readLine());   // from SYSTEM.CNF
    ...
    Console.WriteLn(line.c_str());

A line read out of the disc image is handed to the formatter as the
format itself.  A crafted SYSTEM.CNF containing %s or %n is interpreted
rather than printed - an arbitrary read, and with %n a write - reached
by nothing more than loading the image.  It sits on the malformed-entry
path, so it needs a disc that is already unusual, which is exactly the
kind this thread has been about.  Five more non-literal formats were
the same shape without the attacker: ss.str(), message.c_str(), and
two StdStringFromFormat results already formatted once.

A second std::string_view to %s, in GSTextureReplacements, was the same
segfault as IsoFile::open waiting on a different path.

The rest are type mismatches that print nonsense or read the wrong
width: eight FPRoundMode enums through %d, two affinity masks as %llx
when they are unsigned long, a uptr pair through %p, an s64 through
%d, and two shader selectors against PRIX64.

Verified by deleting the object of every .cpp that includes Console.h
and rebuilding: 42 translation units compiled, zero format warnings.
An incremental build is not evidence here - it silently skips the
files whose warnings you are trying to count, which is how three of
these fixes nearly shipped missing.  The core still loads a 4.6 GiB
image and unloads it twice, exit 0.

The attribute is the point of the commit; the 23 fixes are what it
happened to be holding up.  Anything of this shape added from here on
fails the build instead of the disc.

+++ pcsx2/GS/Renderers/OpenGL/GSDeviceOGL.cpp
@@ -1066,7 +1066,7 @@ static void GSDeviceOGLAppendShaderMacro(std::string& macro, const char* name, i
 
 std::string GSDeviceOGL::GetVSSource(VSSelector sel)
 {
-	Console.WriteLn("Compiling new vertex shader with selector 0x%" PRIX64, sel.key);
+	Console.WriteLn("Compiling new vertex shader with selector 0x%" PRIX64, (uint64_t)sel.key);
 
 	std::string macro;
 	GSDeviceOGLAppendShaderMacro(macro, "VS_FST", sel.fst);
@@ -1081,7 +1081,7 @@ std::string GSDeviceOGL::GetVSSource(VSSelector sel)
 
 std::string GSDeviceOGL::GetPSSource(const PSSelector& sel)
 {
-	Console.WriteLn("Compiling new pixel shader with selector 0x%" PRIX64 "%08X", sel.key_hi, sel.key_lo);
+	Console.WriteLn("Compiling new pixel shader with selector 0x%" PRIX64 "%08X", (uint64_t)sel.key_hi, (unsigned)sel.key_lo);
 
 	std::string macro;
 	GSDeviceOGLAppendShaderMacro(macro, "PS_FST", sel.fst);
